lift availableSlots state to BookingPage
