+++ src/components/form/TimePicker.tsx
@@ -1,7 +1,4 @@
-import { useMemo } from 'react';
-import { Controller, useWatch } from 'react-hook-form';
-import { useBooking } from '../../context/BookingContext';
-import { checkSameDate } from '../../utils/time';
+import { Controller } from 'react-hook-form';
 
 export default function TimePicker({
   timeSlots,
@@ -10,27 +7,6 @@ export default function TimePicker({
   timeSlots: string[];
   name: string;
 }) {
-  const watchedDate = useWatch({ name: 'date' });
-  const { bookingList, isRescheduling } = useBooking();
-
-  const availableSlots = useMemo(() => {
-    if (bookingList.length && watchedDate && !isRescheduling) {
-      // Filter booking list to only include bookings on the selected date
-      const bookingsForSelectedDate = bookingList.filter((booking) =>
-        checkSameDate(booking.date, watchedDate)
-      );
-
-      // Map the booked times for that date
-      const bookedTimes = bookingsForSelectedDate.map(
-        (booking) => booking.time
-      );
-
-      // Filter out the time slots that are already booked
-      return timeSlots.filter((slot) => !bookedTimes.includes(slot));
-    }
-    return timeSlots;
-  }, [bookingList, watchedDate, isRescheduling, timeSlots]);
-
   return (
     <Controller
       name={name}
@@ -38,9 +14,9 @@ export default function TimePicker({
         return (
           <div className="flex flex-col mt-5">
             <h2 className="font-bold mb-4">Please select a time</h2>
-            {availableSlots.length ? (
+            {timeSlots.length ? (
               <>
-                {availableSlots.map((slot, index) => (
+                {timeSlots.map((slot, index) => (
                   <div key={index} className="mb-4">
                     <label className="flex items-center cursor-pointer">
                       <input
