+++ src/pages/BookingPage.tsx
@@ -1,3 +1,4 @@
+import { useMemo } from 'react';
 import { useForm, SubmitHandler, Controller } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
 import { yupResolver } from '@hookform/resolvers/yup';
@@ -8,7 +9,11 @@ import CustomDatePicker from '../components/form/CustomDatePicker';
 import TimePicker from '../components/form/TimePicker';
 import Button from '../components/Button';
 import { bookingSessionForm } from '../utils/validation';
-import { generateTimeSlots, getNearestTimeSlot } from '../utils/time';
+import {
+  checkSameDate,
+  generateTimeSlots,
+  getNearestTimeSlot,
+} from '../utils/time';
 import { useBooking } from '../context/BookingContext';
 const SESSION_LIST: SessionType[] = [
   {
@@ -29,7 +34,8 @@ const SESSION_LIST: SessionType[] = [
 ];
 
 export default function BookingPage() {
-  const { setSelectedSession, selectedSession } = useBooking();
+  const { setSelectedSession, selectedSession, bookingList, isRescheduling } =
+    useBooking();
   const methods = useForm<BookingSessionFormType>({
     resolver: yupResolver(bookingSessionForm),
     defaultValues: selectedSession || {},
@@ -47,6 +53,19 @@ export default function BookingPage() {
     getNearestTimeSlot(watchedDate ? new Date(watchedDate) : new Date()),
     16
   );
+
+  const availableSlots = useMemo(() => {
+    if (bookingList.length && watchedDate && !isRescheduling) {
+      const bookingsForSelectedDate = bookingList.filter((booking) =>
+        checkSameDate(booking.date, watchedDate)
+      );
+      const bookedTimes = bookingsForSelectedDate.map(
+        (booking) => booking.time
+      );
+      return timeSlots.filter((slot) => !bookedTimes.includes(slot));
+    }
+    return timeSlots;
+  }, [bookingList, watchedDate, isRescheduling, timeSlots]);
 
   const onSubmit: SubmitHandler<BookingSessionFormType> = (data) => {
     setSelectedSession(data);
@@ -87,7 +106,7 @@ export default function BookingPage() {
             }}
           />
           <CustomDatePicker name="date" />
-          <TimePicker name="time" timeSlots={timeSlots} />
+          <TimePicker name="time" timeSlots={availableSlots} />
           {watchedDate && watchedSession && watchedTime && (
             <Button title="Continue" type="submit" />
           )}
